requirements.txt
aiofiles~=22.1.0
d20~=1.1.2
httpx[http2]~=0.23.2
jq~=1.3.0
lark~=1.1.5
Pillow~=9.4.0
Pyrogram==2.0.93
python-magic~=0.4.27
redis~=4.4.0
TgCrypto~=1.2.5
